Relocated `CONTENTMENT_LIQUID_CONTEXT` constant
to the same file as the context code itself.
Reduces an extra import.
