+++ src/Umbraco.Community.Contentment.Client/src/global-context/index.ts
@@ -1,5 +1,5 @@
 // SPDX-License-Identifier: MIT
 // Copyright © 2025 Lee Kelleher
 
-export { CONTENTMENT_LIQUID_CONTEXT } from './liquid/liquid.context-token.js';
+export { CONTENTMENT_LIQUID_CONTEXT } from './liquid/liquid.context.js';
 export type * from './liquid/liquid.context.js';
