+++ src/Umbraco.Community.Contentment.Client/src/global-context/liquid/index.ts
@@ -1,5 +1,5 @@
 // SPDX-License-Identifier: MIT
 // Copyright © 2025 Lee Kelleher
 
-export { CONTENTMENT_LIQUID_CONTEXT } from './liquid.context-token.js';
+export { CONTENTMENT_LIQUID_CONTEXT } from './liquid.context.js';
 export type { ContentmentLiquidContext } from './liquid.context.js';
