--- src/Umbraco.Community.Contentment.Client/src/global-context/liquid/liquid.context-token.ts
@@ -1,7 +1,0 @@
-// SPDX-License-Identifier: MIT
-// Copyright © 2025 Lee Kelleher
-
-import { UmbContextToken } from '@umbraco-cms/backoffice/context-api';
-import type { ContentmentLiquidContext } from './liquid.context.js';
-
-export const CONTENTMENT_LIQUID_CONTEXT = new UmbContextToken<ContentmentLiquidContext>('ContentmentLiquidContext');
